refactored files
refactored files that are responsible for api integration to data package
