+++ app/src/main/java/com/example/zadumite_frontend/data/authenticator/AuthAuthenticator.kt
@@ -1,4 +1,4 @@
-package com.example.zadumite_frontend.network.authenticator
+package com.example.zadumite_frontend.data.authenticator
 
 import com.example.zadumite_frontend.data.model.token.JwtTokenManager
 import kotlinx.coroutines.runBlocking
