+++ app/src/main/java/com/example/zadumite_frontend/di/InterceptorModule.kt
@@ -1,10 +1,10 @@
 package com.example.zadumite_frontend.di
 
-import com.example.zadumite_frontend.network.interception.AccessTokenInterceptor
-import com.example.zadumite_frontend.network.interception.RefreshTokenInterceptor
+import com.example.zadumite_frontend.data.interception.AccessTokenInterceptor
+import com.example.zadumite_frontend.data.interception.RefreshTokenInterceptor
 import org.koin.dsl.module
 
 val interceptorModule = module {
     single { AccessTokenInterceptor(get()) }
-    single { RefreshTokenInterceptor(get())}
+    single { RefreshTokenInterceptor(get()) }
 }
