+++ app/src/main/java/com/example/zadumite_frontend/di/providers/provideAccessOkHttpClient.kt
@@ -1,7 +1,7 @@
 package com.example.zadumite_frontend.di.providers
 
-import com.example.zadumite_frontend.network.authenticator.AuthAuthenticator
-import com.example.zadumite_frontend.network.interception.AccessTokenInterceptor
+import com.example.zadumite_frontend.data.authenticator.AuthAuthenticator
+import com.example.zadumite_frontend.data.interception.AccessTokenInterceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import java.util.concurrent.TimeUnit
