+++ app/src/main/java/com/example/zadumite_frontend/di/providers/provideRefreshOkHttpClient.kt
@@ -1,6 +1,6 @@
 package com.example.zadumite_frontend.di.providers
 
-import com.example.zadumite_frontend.network.interception.RefreshTokenInterceptor
+import com.example.zadumite_frontend.data.interception.RefreshTokenInterceptor
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import java.util.concurrent.TimeUnit
