+++ app/src/main/java/com/example/zadumite_frontend/domain/FetchWordUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.zadumite_frontend.domain
 
 import com.example.zadumite_frontend.data.model.word.Word
-import com.example.zadumite_frontend.network.ZaDumiteApiService
+import com.example.zadumite_frontend.data.api.ZaDumiteApiService
 
 class FetchWordUseCase(private val apiService : ZaDumiteApiService) {
     suspend operator fun invoke(): Word {
